+++ src/components/modules/tools.ts
@@ -123,6 +123,7 @@ export default class Tools extends Module {
 
   /**
    * Constant for available Tools Settings
+   * @todo separate internal and external options
    * @return {object}
    */
   public get apiSettings() {
@@ -135,6 +136,7 @@ export default class Tools extends Module {
       SHORTCUT: 'shortcut',
       TOOLBOX: 'toolbox',
       SANITIZE_CONFIG: 'sanitize',
+      CONVERSION_CONFIG: 'conversionConfig',
     };
   }
 
@@ -151,7 +153,7 @@ export default class Tools extends Module {
   private readonly toolsAvailable: {[name: string]: ToolConstructable} = {};
 
   /**
-   * Tools` classes not availbale to use beacause of preparation failure
+   * Tools` classes not available to use because of preparation failure
    */
   private readonly toolsUnavailable: {[name: string]: ToolConstructable} = {};
 
@@ -207,7 +209,7 @@ export default class Tools extends Module {
     /**
      * Assign internal tools
      */
-    _.deepMerge(this.config.tools, this.internalTools);
+    this.config.tools = _.deepMerge({}, this.internalTools, this.config.tools);
 
     if (!this.config.hasOwnProperty('tools') || Object.keys(this.config.tools).length === 0) {
       throw Error('Can\'t start without tools');
@@ -302,14 +304,19 @@ export default class Tools extends Module {
     /**
      * Configuration to be passed to the Tool's constructor
      */
-    const config = this.toolsSettings[tool][this.apiSettings.CONFIG];
+    const config = this.toolsSettings[tool][this.apiSettings.CONFIG] || {};
+
+    // Pass placeholder to initial Block config
+    if (tool === this.config.initialBlock && !config.placeholder) {
+      config.placeholder = this.config.placeholder;
+    }
 
     /**
      * @type {{api: API, config: ({}), data: BlockToolData}}
      */
     const constructorOptions = {
       api: this.Editor.API.methods,
-      config: config || {},
+      config,
       data,
     };